web/requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
pillow>=10.0.0
google-generativeai>=0.3.0
plotly>=5.18.0
PyYAML>=6.0.0
python-dotenv>=1.0.0